Omptimize guiscrcpy
Guiscrcpy had been calling functions to calculate screen_size every now and then,
However this is not necessary, as it needs to be called only once.

(ii) guiscrcpy has swipes which were made to be different functions. The code was optimized by making the function calls to a single function and add args to it

(iii) Remove redundant comments
Some code was commented out for future refernce was not needed any later