fix: Fix scrcpy-server path being null if not set in config (Windows)
Fixes #172
Scrcpy server path was set as null in the config and guiscrcpy did not try to ask the user
for the path to scrcpy-server despite being null. This was because of a missing condition which
checks if scrcpy-server is null. Now it shows a file dialog box if scrcpy-server is Null